cluster/dht: Refactor dht-inode-write fops
Removed code inconsistencies, duplication in fop_cbk(), fop2() for all
inode-write operations.

+++ xlators/debug/delay-gen/src/delay-gen.h
@@ -21,7 +21,7 @@ typedef struct {
     int enable[GF_FOP_MAXVALUE];
     int op_count;
     int delay_ppm;
-    int delay_duration;
+    uint64_t delay_duration;
 } dg_t;
 
 #endif /* __DELAY_GEN_H__ */
